+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - `project add` fails when there is no project
+- colors breaking alignment of tabwriter
 
 ## [0.5.1] - 2017-09-15
 
@@ -27,9 +28,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - `resize` message output
-
-### Removed
-
 
 ## [0.5.0] - 2017-09-14