+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- release linux and darwin as `tar.gz`
+
 ### Fixed
 
 - `config unset` accepts `--team` flag