+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- release linux and darwin as `tar.gz`
+
+### Fixed
+
+- `config unset` accepts `--team` flag
+- `config unset` fails if no key is passed
+- `config unset` fails with a friendlier message
+
 ## [0.6.0] - 2017-10-03
 
 ### Added